Babel setup for ES6 completed.
