+++ main.js
@@ -1,0 +1,5 @@
+require('babel-register')({
+    presets: ['env']
+});
+
+module.exports = require('./test.js')
