+++ src/classes/player.js
@@ -1,0 +1,23 @@
+//Class and Constructor to initiate player object.
+class player {
+    constructor(currency) {
+        this.currency = currency;
+    }
+
+    //Function to return currency.
+    getCurrency() {
+        return this.currency;
+    }
+
+    //Function to add money.
+    gainMoney(money) {
+        this.currency += money;
+    }
+
+    //Function to subtract money.
+    spendMoney(money) {
+        this.currency -= money;
+    }
+}
+
+export { player };
