+++ test.js
@@ -1,0 +1,7 @@
+import { player } from './src/classes/player';
+import { monster } from './src/classes/monster';
+import { tower } from './src/classes/tower';
+
+let user = new player(100);
+
+console.log(user.getCurrency());
